extract identity
